# Conflicts:
#	resources/js/Shared/Sidebar.vue

+++ app/Http/Controllers/web/partner/ProgramsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\web\partner;
 use App\Http\Controllers\web\WebBaseController;
 use App\Http\Requests\BriefcaseRequest;
 use App\Models\Briefcase;
+use App\Models\BriefcaseChange;
 use App\Models\BriefcaseType;
 use App\Models\Purchase;
 use App\Models\User;
@@ -32,6 +33,7 @@ class ProgramsController extends WebBaseController
     public function index()
     {
         $briefcases = Briefcase::where('partner_id', Auth::user()->id)
+            ->with('active_change')
             ->get();
 
         return Inertia::render('Programs/Index', [
@@ -64,7 +66,6 @@ class ProgramsController extends WebBaseController
     public function update(BriefcaseRequest $request)
     {
         $data = $request->all();
-
         if ($request->has('image')) {
             $filepath = $this->attachmentService->storeFile(
                 $request->file('image'),
@@ -76,16 +77,37 @@ class ProgramsController extends WebBaseController
 
         $data['partner_id'] = Auth::user()->id;
 
-        $briefcase = $this->briefcaseService->update(
-            $request->input('id'),
-            $data
-        );
+        if (Auth::user()->isAdmin) {
+            $briefcase = $this->briefcaseService->update(
+                $request->input('id'),
+                $data
+            );
 
-        if (!empty($briefcase)) {
-            return redirect()
-                ->route('programs-crud.index');
-        } else {
-            return $this->responseFail('Не удалось обновить портфель');
+            if (!empty($briefcase)) {
+                return redirect()
+                    ->route('programs-crud.index');
+            } else {
+                return $this->responseFail('Не удалось обновить портфель');
+            }
+        } elseif (Auth::user()->isPartner) {
+            $change = BriefcaseChange::create([
+                'briefcase_id' => data_get($request, 'id'),
+                'partner_id' => Auth::user()->id,
+                'status' => BriefcaseChange::STATUS_CREATED,
+                'type_id' => BriefcaseChange::TYPE_EDIT,
+                'change_data' => $data
+            ]);
+
+            if (!empty($change)) {
+                return redirect()
+                    ->back()
+                    ->with([
+                        'message' => 'Запрос на изменение программы успешно отправлен',
+                        'sub_message' => 'Ожидайте ответа модератора'
+                    ]);
+            } else {
+                return $this->responseFail('Не удалось создать завпрос');
+            }
         }
     }
 
@@ -101,25 +123,55 @@ class ProgramsController extends WebBaseController
         $data['image_path'] = $filepath;
 
         $data['partner_id'] = Auth::user()->id;
-
         $briefcase = $this->briefcaseService->create($data);
-        if (!empty($briefcase)) {
-            return redirect()
-                ->route('programs-crud.index');
-        } else {
-            return $this->responseFail('failed saving briefcase');
-        }
+        $briefcase->delete();
+
+        BriefcaseChange::create([
+            'briefcase_id' => data_get($briefcase, 'id'),
+            'partner_id' => Auth::user()->id,
+            'status' => BriefcaseChange::STATUS_CREATED,
+            'type_id' => BriefcaseChange::TYPE_CREATE,
+            'data' => null
+        ]);
+
+        return redirect()
+            ->route('programs-crud.index')
+            ->with([
+                'message' => 'Запрос на добавление программы успешно отправлен',
+                'sub_message' => 'Ожидайте ответа модератора'
+            ]);
     }
 
     public function destroy($id)
     {
-        $deleted = $this->briefcaseService->delete($id);
+        if (Auth::user()->isAdmin) {
+            $deleted = $this->briefcaseService->delete($id);
 
-        if ($deleted) {
-            return redirect()
-                ->route('programs-crud.index');
-        } else {
-            return $this->responseFail('failed deleting briefcase');
+            if ($deleted) {
+                return redirect()
+                    ->route('programs-crud.index');
+            } else {
+                return $this->responseFail('failed deleting briefcase');
+            }
+        } elseif (Auth::user()->isPartner) {
+            $change = BriefcaseChange::create([
+                'briefcase_id' => $id,
+                'partner_id' => Auth::user()->id,
+                'status' => BriefcaseChange::STATUS_CREATED,
+                'type_id' => BriefcaseChange::TYPE_DELETE,
+                'data' => null
+            ]);
+
+            if (!empty($change)) {
+                return redirect()
+                    ->back()
+                    ->with([
+                        'message' => 'Запрос на удаление программы успешно отправлен',
+                        'sub_message' => 'Ожидайте ответа модератора'
+                    ]);
+            } else {
+                return $this->responseFail('Не удалось создать запрос');
+            }
         }
     }
 }
